+++ packages/core-components/src/lib/unsorted/viz/core/SearchBar.svelte
