+++ packages/core-components/src/lib/unsorted/viz/index.js
@@ -1,28 +1,28 @@
-export { default as AreaChart } from './AreaChart.svelte';
-export { default as Area } from './Area.svelte';
-export { default as BarChart } from './BarChart.svelte';
-export { default as Bar } from './Bar.svelte';
-export { default as BigValue } from './BigValue.svelte';
-export { default as BoxPlot } from './BoxPlot.svelte';
-export { default as BubbleChart } from './BubbleChart.svelte';
-export { default as Bubble } from './Bubble.svelte';
-export { default as Chart } from './Chart.svelte';
-export { default as Column } from './Column.svelte';
-export { default as DataTable } from './DataTable.svelte';
-export { default as EchartsCopyTarget } from './EchartsCopyTarget.svelte';
-export { default as EChartsMap } from './EChartsMap.svelte';
-export { default as ECharts } from './ECharts.svelte';
-export { default as ErrorChart } from './ErrorChart.svelte';
-export { default as FunnelChart } from './FunnelChart.svelte';
-export { default as Histogram } from './Histogram.svelte';
-export { default as Hist } from './Hist.svelte';
-export { default as LineChart } from './LineChart.svelte';
-export { default as Line } from './Line.svelte';
-export { default as SankeyChart } from './SankeyChart.svelte';
-export { default as ScatterPlot } from './ScatterPlot.svelte';
-export { default as Scatter } from './Scatter.svelte';
-export { default as SearchBar } from './SearchBar.svelte';
-export { default as ReferenceLine } from './ReferenceLine.svelte';
-export { default as ReferenceArea } from './ReferenceArea.svelte';
-export { default as USMap } from './USMap.svelte';
-export { default as Value } from './Value.svelte';
+export { default as AreaChart } from './area/AreaChart.svelte';
+export { default as Area } from './area/Area.svelte';
+export { default as BarChart } from './bar/BarChart.svelte';
+export { default as Bar } from './bar/Bar.svelte';
+export { default as BigValue } from './core/BigValue.svelte';
+export { default as BoxPlot } from './box/BoxPlot.svelte';
+export { default as BubbleChart } from './bubble/BubbleChart.svelte';
+export { default as Bubble } from './bubble/Bubble.svelte';
+export { default as Chart } from './core/Chart.svelte';
+export { default as Column } from './table/Column.svelte';
+export { default as DataTable } from './table/DataTable.svelte';
+export { default as EChartsCopyTarget } from './core/EChartsCopyTarget.svelte';
+export { default as EChartsMap } from './map/EChartsMap.svelte';
+export { default as ECharts } from './core/ECharts.svelte';
+export { default as ErrorChart } from './core/ErrorChart.svelte';
+export { default as FunnelChart } from './funnel/FunnelChart.svelte';
+export { default as Histogram } from './histogram/Histogram.svelte';
+export { default as Hist } from './histogram/Hist.svelte';
+export { default as LineChart } from './line/LineChart.svelte';
+export { default as Line } from './line/Line.svelte';
+export { default as SankeyChart } from './sankey/SankeyChart.svelte';
+export { default as ScatterPlot } from './scatter/ScatterPlot.svelte';
+export { default as Scatter } from './scatter/Scatter.svelte';
+export { default as SearchBar } from './core/SearchBar.svelte';
+export { default as ReferenceLine } from './references/ReferenceLine.svelte';
+export { default as ReferenceArea } from './references/ReferenceArea.svelte';
+export { default as USMap } from './map/USMap.svelte';
+export { default as Value } from './core/Value.svelte';
